circuits.web.wsgi: Fixed WSGI Application deployment of circuits.web web applications.
New review
Labels: Priority-High Milestone-1.2.1

Ensure that this is the right way to go about fixing circutis.web application deployment via the Application component. Without flushing events like this, any Controller(s) in the system are not registered with the Dispatcher as the system is essentially lifeless in a non-running state.