Major non-breaking / incompatible change wrt to pollers.
NOTE: This is a library layout change that affects imports of pollers.

Old:
from circuits.net.pollers import ...

NEW:
from circuits.core.pollers import ...

Also added:

The Manager and any Component when run (either as a thread, process or as the main process) will attempt to "locate" a poller in the system. If one is not found, it will instantiate one and register it. This ensures that a system only ever has *one* single poller that all I/O and Socket components use.

Also added:

A new example (examples/udplatency.py) which demonstrates measuring the
"round trip" latecny of UDP packets from client to server and back to client
using the circuits framework.