- circuits.net: Fixed ``Server.host`` and ``Server.port`` properties
The ``Server.host`` property will reteurn what ``getsockname()`` returns
on the underlying listening socket and return it's first item if it's a
tuple, otherwise it will return the entire string (eg: a UNIX Socket).

The ``Server.port`` property does a similar thing but returns ``None``
in the case of ``getsockname()`` **not** returning a tuple.

- circuits.net: Fixed a few missing parameters to ``Server``