New Feature: Pass the failed event as an optional kearg for ``Error`` events.
Anytime an exception is raised that isn't handled, an optinoal kwarg
``fevent`` (_Failed Event_) is passed along with the associated ``Error``
event triggered by the exception.

This allows a developer to deal with unexpected behavior by inspecting the
original event and the data it was carrying. For example, we now use this
to determine if an error occured in circuits.web before a proper Request/response
set of objects were able to be created. If such an error occured, we capture this
via an ``error`` event handler and inspect ``fevent`` extracting the client socket
in question and send an error response.

Handling ``fevent`` looks like this:

.. code-block:: python
    :linenos:

    @handler("error")
    def _on_error(self, etype, evalue, etraceback, handler=None, fevent=None):
       pass