Fix error handling if wrapping TLS socket fails
Wrapping the SSL socket might go wrong for various reasons.
For example if the private key is not readable by the user or does not
exist. In this case we should not let the connection opened forever but
make sure it is closed. Even better would be to check at socket startup
if everything is okay (maybe in the future?).