Format tracebacks in exception event.
The traceback argument for the exception event was formatted in every
place where it was instantiated. Removed this code redundancy by
oursourcing this into the exception event class. The current behavior
also does not fit the docstring of the exception event class as it
expects a unformatted traceback as parameter. This handling is now
better adaptable and in the future we might remove the formatation
completely.

Some more cleanup has been done:
* Manager._values is nowhere used
* PEP8 says don't do if len(some_list)
* remove duplicated str() casting