Fix crash from deleting undefined variables
This address a bug introduced in 3.2 that occurs when running a Circuits
app with packaged with py2exe into a single zip file. In that case, the
innermost if-check would never pass, the fd and exec_ variables would never be
defined, and the final del would raise an unhandled NameError.