Documentation: Cannot import name `WebSockets` (#230)
I think the example code for Websockets might be slightly out of date.

I attempted to run the code, and python returned an error:
```
Traceback (most recent call last):
  File "file.py", line 5, in <module>
    from circuits.web.dispatchers import WebSockets
ImportError: cannot import name 'WebSockets'
```
I believe it's because it should be importing `WebSocketsDispatcher` instead. The examples in the examples directory would seem to agree with this.