Added a bufsize arguement to the init of BaseServer
The use case for this is using a class that inherits from BaseServer such as Server and I would like to expose BUFSIZE as a configurable value.
Currently in circuits.net.sockets BUFSIZE is a constant of 4096 and this value is used unless set otherwise in the init of something like TCPServer.

You can import CONSTANTS, so using the already existing BUFSIZE CONSTANT as the default value if a user doesn't set bufsize themselves